Npm linux build script (#8)
* Fix compiling issue under Linux

* add package definition for az-iot-gw-win and az-iot-gw-lin.

* Refactor the package definition and start the shell script for building package on linux.

* Build script for linux distribution of iot gw.

* Fix the npm pack exit issue.

* add package distribution files into the ignore list and related main.js files.

* Fix the folder creation issue.

* add the dependency for az-iot-gw to support linux.

* Fix the warning and loading path issue.

* use static link.

* Change the default rpath of gw executable.

* set the rpath for gw.

* Setup RPATH for executable file.

* Disable the RPATH.

* Fix the LD_LIBRARY_PATH issue.

* Revert changes on CMakeLists.txt

+++ dist_pkgs/gw/src/main.c
@@ -25,7 +25,7 @@ bool validate_args(int argc, char* argv[])
     return false;
   }
 
-  size_t i = 1;
+  int i = 1;
   while (i < EXPECTED_PARAMS && strlen(argv[i]) != 0) {
     ++i;
   }
